rigol-dg: add support for AWG of MSO7000 series
The MSO7000 has two arbitrary waveform generators built in. The
command set is very similar to the one used by the dedicated Rigol
AWGs but not quite the same.

The responses to the "APPL?" command return the short form of the
waveform name so we compare only the mandatory parts of the name and
ignore the remaining characters.

The MSO7000 does not have a dedicated frequency counter input; instead
any of the four oscilloscope channels can use the "frequency counter"
measurement option. We deactivate frequency counter support in
rigol-dg for the MSO7000 series as it's related to the oscilloscope
channels which are managed by rigol-ds.

Support for the following features can be added at a later time:
- waveforms Sinc, ExpRise, ExpFall, ECG, Gauss, Lorentz, Haversine
- setting Ramp symmetry percentage
- setting output impedance
- configuring modulation