bindings/python: fix exception handling
The documentation of the Python bindings claimed that the
sigrok.core.Error exception class is used to indicate libsigrok
errors. In practice it raised RuntimeError instead most of the
time. This is not just a deviation from the documentation (which could
be adjusted to match the implementation) but has a serious impact on
applications using the sigrok Python bindings: RuntimeError is derived
from BaseException rather than Exception so it won't usually be caught
by exception handlers. RuntimeError is meant for serious problems and
normally causes the whole program to abort immediately. Catching it is
considered bad practice.

Fix this by setting up a dedicated exception handler for Python that
creates and raises a copy of the sigrok::Error instance we caught. For
developer convenience we add an __str__() implementation (used when
the exception is converted to a string) that shows both the error
number and the explanation string (Error.what()).

It's worth considering whether the other bindings should also use
sigrok::Error but I'm not sure what the impact of this API change
would be for the other languages. In Python it's straightforward to
support both old (RuntimeError) and new (sigrok.core.Error) exception
classes; the same might not be true in other languages.