rigol-dg: add support for output impedance
All product series currently supported by rigol-dg support setting the
"output impedance". Or rather setting the expected load impedance —
the setting does not actually alter the behaviour of the output itself
in hardware; instead it scales the amplitude so that if a load of the
configured impedance is actually connected the amplitude measured
across the load matches the configured amplitude.

The stand-alone AWGs from the DG800, DG900 and DG1000Z series support
HighZ (encoded as "INF" when writing and a very large number when
reading) and integer values from 1Ω to 10kΩ. The AWGs built into the
MSO7000 series oscilloscopes support HighZ (encoded as "OMEG") and
50Ω (encoded as "FIFT").